Updated mingw compilation
git-svn-id: https://svn.pjsip.org/repos/pjproject/trunk@241 74dad513-b988-da41-8d7b-12977e46ad98

+++ pjsip/src/pjsua-lib/pjsua_call.c
@@ -853,7 +853,7 @@ static void pjsua_call_on_media_update(pjsip_inv_session *inv,
 	    len = pj_ansi_sprintf( info+info_len,
 				   ", stream #%d: %.*s (%s)", i,
 				   (int)strm_info->fmt.encoding_name.slen,
-				   (int)strm_info->fmt.encoding_name.ptr,
+				   strm_info->fmt.encoding_name.ptr,
 				   dir);
 	    if (len > 0)
 		info_len += len;
